requirements.txt
pyTelegramBotAPI == 4.11.0
requests == 2.28.2
telebot-calendar == 1.2
python-dotenv == 1.0.0